# [8_4] improve error handling and performance of mimalloc