--- a/tests/Kernel/Types/string_test.cpp
+++ b/tests/Kernel/Types/string_test.cpp
@@ -1,6 +1,5 @@
 #include "a_lolly_test.hpp"
 #include "string.hpp"
-#define ANKERL_NANOBENCH_IMPLEMENT
 #include <nanobench.h>
 
 static ankerl::nanobench::Bench bench;