--- a/tests/System/Classes/file_url_test.cpp
+++ b/tests/System/Classes/file_url_test.cpp
@@ -4,7 +4,7 @@
  *  \author Darcy
  *  \date   2019-2023
  */
-#include "a_tbox_main.cpp"
+#include "a_lolly_test.hpp"
 #include "url.hpp"
 
 #if defined(OS_MINGW) || defined(OS_WIN)