--- a/tests/System/Files/file_test.cpp
+++ b/tests/System/Files/file_test.cpp
@@ -4,8 +4,9 @@
  *  \author Darcy
  *  \date   2019-2023
  */
-#include "a_tbox_main.cpp"
+#include "a_lolly_test.hpp"
 #include "file.hpp"
+#include "sys_utils.hpp"
 #include "tbox/tbox.h"
 
 url unix_root= url_system ("/");