--- a/tests/a_lolly_test.hpp
+++ b/tests/a_lolly_test.hpp
@@ -1,4 +1,2 @@
-#define DOCTEST_CONFIG_IMPLEMENT_WITH_MAIN
-
 #include "doctest/doctest.h"
 #include "lolly_doctests.hpp"