--- a/tests/lolly/data/unicode_test.cpp
+++ b/tests/lolly/data/unicode_test.cpp
@@ -9,7 +9,7 @@
  * in the root directory or <http://www.gnu.org/licenses/gpl-3.0.html>.
  ******************************************************************************/
 
-#include "a_tbox_main.cpp"
+#include "a_lolly_test.hpp"
 #include "lolly/data/unicode.hpp"
 
 using lolly::data::has_cjk_unified_ideographs;