by register error handler to avoid checking NULL pointer at every allocation. Thus lolly must be initialized by `init_tbox` before any invocation. Therefore many tests are rewritten.

Such optimization cannot be performed on jemalloc or fast_alloc.