@darcy-shen 2024-01-18
I do not think it is a perf pr, but it is a pr for better error handling.