@darcy-shen 2024-01-18
You can disable wasm CI on windows, it is better to do wasm development on linux. 

I think wasm dev env on Windows is nice to have but it is not necessary. What's your idea on wasm dev env, if you agree my idea, just disable the ci in this pr.