@darcy-shen 2024-01-18
> I do not think it is a perf pr

Could you provide the benchmark result? Null check here is not the bottle neck of performance here.

>  but it is a pr for better error handling.

For error handling, this pr is important for developer to finding bugs for the memory allocator.